solution/requirements.txt
pandas
duckdb
plotly
flask
python-dotenv
pyspark
sqlalchemy